requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
cattrs==24.1.3
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
distro==1.9.0
exceptiongroup==1.3.0
flatbuffers==25.2.10
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.25.0
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
gTTS==2.5.4
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
jh2==5.0.9
Jinja2==3.1.6
jiter==0.10.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
narwhals==1.40.0
niquests==3.14.1
numpy==2.0.2
oauthlib==3.2.2
openai==1.84.0
opencv-python==4.11.0.86
openmeteo_requests==1.5.0
openmeteo_sdk==1.20.0
packaging==24.2
pandas==2.2.3
pillow==11.2.1
platformdirs==4.3.8
proto-plus==1.26.1
protobuf==6.31.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pydeck==0.9.1
pydub==0.25.1
pygame==2.6.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
qh3==1.5.1
redis==6.1.0
referencing==0.36.2
requests==2.32.3
requests-cache==1.2.1
requests-oauthlib==2.0.0
retry-requests==2.0.0
rpds-py==0.25.1
rsa==4.9.1
scipy==1.13.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.2
spotipy==2.25.1
streamlit==1.45.1
streamlit-autorefresh==1.0.1
streamlit-camera-input-live==0.2.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.2.0
url-normalize==2.2.1
urllib3==2.4.0
urllib3-future==2.12.922
wassima==1.2.2
